5/6

+++ src/main/java/com/ohgiraffers/practicespringdatajpa/tour/controller/TourController.java
@@ -66,6 +66,9 @@ public class TourController {
     @GetMapping("/delete")
     public String delete(@RequestParam("no") String no) {
         if (no != null && !"".equals(no)) {
+            if(Integer.parseInt(no) > 206 || Integer.parseInt(no) < 0 ) {
+                return "tour/delete";
+            }
             tourService.deleteTourInfo(no);
             return "redirect:/tour/allSelect";
         }
